Migrate native_set
- Migrate native_set to new implementation of the native type binding api
