+++ abra_core/src/builtins/gen_native_types.rs
@@ -531,306 +531,3 @@ impl NativeType for NativeString {
         vec![]
     }
 }
-pub struct NativeSet;
-pub trait NativeSetMethodsAndFields {
-    fn field_size(obj: Box<Value>) -> Value;
-    fn method_to_string(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_is_empty(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_enumerate(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_contains(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_insert(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_remove(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_map(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_filter(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_reduce(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_as_array(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_union(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_difference(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM) -> Option<Value>;
-    fn method_intersection(receiver: Option<Value>, args: Vec<Value>, vm: &mut VM)
-        -> Option<Value>;
-}
-impl NativeType for NativeSet {
-    fn get_field_type(name: &str) -> Option<(usize, Type)> {
-        match name {
-            "size" => Some((0usize, Type::Int)),
-            _ => None,
-        }
-    }
-    fn get_method_type(name: &str) -> Option<(usize, Type)> {
-        match name {
-            "toString" => Some((
-                0usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![],
-                    ret_type: Box::new(Type::String),
-                    is_variadic: false,
-                }),
-            )),
-            "isEmpty" => Some((
-                1usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![],
-                    ret_type: Box::new(Type::Bool),
-                    is_variadic: false,
-                }),
-            )),
-            "enumerate" => Some((
-                2usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![],
-                    ret_type: Box::new(Type::Array(Box::new(Type::Tuple(vec![
-                        Type::Generic("T".to_string()),
-                        Type::Int,
-                    ])))),
-                    is_variadic: false,
-                }),
-            )),
-            "contains" => Some((
-                3usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![("value".to_string(), Type::Generic("T".to_string()), false)],
-                    ret_type: Box::new(Type::Bool),
-                    is_variadic: false,
-                }),
-            )),
-            "insert" => Some((
-                4usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![("value".to_string(), Type::Generic("T".to_string()), false)],
-                    ret_type: Box::new(Type::Unit),
-                    is_variadic: false,
-                }),
-            )),
-            "remove" => Some((
-                5usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![("value".to_string(), Type::Generic("T".to_string()), false)],
-                    ret_type: Box::new(Type::Unit),
-                    is_variadic: false,
-                }),
-            )),
-            "map" => Some((
-                6usize,
-                Type::Fn(FnType {
-                    type_args: vec!["U".to_string()],
-                    arg_types: vec![(
-                        "fn".to_string(),
-                        Type::Fn(FnType {
-                            type_args: vec![],
-                            arg_types: vec![(
-                                "_".to_string(),
-                                Type::Generic("T".to_string()),
-                                false,
-                            )],
-                            ret_type: Box::new(Type::Generic("U".to_string())),
-                            is_variadic: false,
-                        }),
-                        false,
-                    )],
-                    ret_type: Box::new(Type::Array(Box::new(Type::Generic("U".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            "filter" => Some((
-                7usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![(
-                        "fn".to_string(),
-                        Type::Fn(FnType {
-                            type_args: vec![],
-                            arg_types: vec![(
-                                "_".to_string(),
-                                Type::Generic("T".to_string()),
-                                false,
-                            )],
-                            ret_type: Box::new(Type::Bool),
-                            is_variadic: false,
-                        }),
-                        false,
-                    )],
-                    ret_type: Box::new(Type::Set(Box::new(Type::Generic("T".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            "reduce" => Some((
-                8usize,
-                Type::Fn(FnType {
-                    type_args: vec!["U".to_string()],
-                    arg_types: vec![
-                        (
-                            "initialValue".to_string(),
-                            Type::Generic("U".to_string()),
-                            false,
-                        ),
-                        (
-                            "fn".to_string(),
-                            Type::Fn(FnType {
-                                type_args: vec![],
-                                arg_types: vec![
-                                    ("_".to_string(), Type::Generic("U".to_string()), false),
-                                    ("_".to_string(), Type::Generic("T".to_string()), false),
-                                ],
-                                ret_type: Box::new(Type::Generic("U".to_string())),
-                                is_variadic: false,
-                            }),
-                            false,
-                        ),
-                    ],
-                    ret_type: Box::new(Type::Generic("U".to_string())),
-                    is_variadic: false,
-                }),
-            )),
-            "asArray" => Some((
-                9usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![],
-                    ret_type: Box::new(Type::Array(Box::new(Type::Generic("T".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            "union" => Some((
-                10usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![(
-                        "other".to_string(),
-                        Type::Set(Box::new(Type::Generic("T".to_string()))),
-                        false,
-                    )],
-                    ret_type: Box::new(Type::Set(Box::new(Type::Generic("T".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            "difference" => Some((
-                11usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![(
-                        "other".to_string(),
-                        Type::Set(Box::new(Type::Generic("T".to_string()))),
-                        false,
-                    )],
-                    ret_type: Box::new(Type::Set(Box::new(Type::Generic("T".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            "intersection" => Some((
-                12usize,
-                Type::Fn(FnType {
-                    type_args: vec![],
-                    arg_types: vec![(
-                        "other".to_string(),
-                        Type::Set(Box::new(Type::Generic("T".to_string()))),
-                        false,
-                    )],
-                    ret_type: Box::new(Type::Set(Box::new(Type::Generic("T".to_string())))),
-                    is_variadic: false,
-                }),
-            )),
-            _ => None,
-        }
-    }
-    fn get_static_field_or_method(_name: &str) -> Option<(usize, Type)> {
-        None
-    }
-    fn get_field_value(obj: Box<Value>, field_idx: usize) -> Value {
-        match field_idx {
-            0usize => Self::field_size(obj),
-            _ => unreachable!(),
-        }
-    }
-    fn get_method_value(obj: Box<Value>, method_idx: usize) -> Value {
-        match method_idx {
-            0usize => Value::NativeFn(NativeFn {
-                name: "toString",
-                receiver: Some(obj),
-                native_fn: Self::method_to_string,
-                has_return: true,
-            }),
-            1usize => Value::NativeFn(NativeFn {
-                name: "isEmpty",
-                receiver: Some(obj),
-                native_fn: Self::method_is_empty,
-                has_return: true,
-            }),
-            2usize => Value::NativeFn(NativeFn {
-                name: "enumerate",
-                receiver: Some(obj),
-                native_fn: Self::method_enumerate,
-                has_return: true,
-            }),
-            3usize => Value::NativeFn(NativeFn {
-                name: "contains",
-                receiver: Some(obj),
-                native_fn: Self::method_contains,
-                has_return: true,
-            }),
-            4usize => Value::NativeFn(NativeFn {
-                name: "insert",
-                receiver: Some(obj),
-                native_fn: Self::method_insert,
-                has_return: false,
-            }),
-            5usize => Value::NativeFn(NativeFn {
-                name: "remove",
-                receiver: Some(obj),
-                native_fn: Self::method_remove,
-                has_return: false,
-            }),
-            6usize => Value::NativeFn(NativeFn {
-                name: "map",
-                receiver: Some(obj),
-                native_fn: Self::method_map,
-                has_return: true,
-            }),
-            7usize => Value::NativeFn(NativeFn {
-                name: "filter",
-                receiver: Some(obj),
-                native_fn: Self::method_filter,
-                has_return: true,
-            }),
-            8usize => Value::NativeFn(NativeFn {
-                name: "reduce",
-                receiver: Some(obj),
-                native_fn: Self::method_reduce,
-                has_return: true,
-            }),
-            9usize => Value::NativeFn(NativeFn {
-                name: "asArray",
-                receiver: Some(obj),
-                native_fn: Self::method_as_array,
-                has_return: true,
-            }),
-            10usize => Value::NativeFn(NativeFn {
-                name: "union",
-                receiver: Some(obj),
-                native_fn: Self::method_union,
-                has_return: true,
-            }),
-            11usize => Value::NativeFn(NativeFn {
-                name: "difference",
-                receiver: Some(obj),
-                native_fn: Self::method_difference,
-                has_return: true,
-            }),
-            12usize => Value::NativeFn(NativeFn {
-                name: "intersection",
-                receiver: Some(obj),
-                native_fn: Self::method_intersection,
-                has_return: true,
-            }),
-            _ => unreachable!(),
-        }
-    }
-    fn get_static_field_values() -> Vec<(String, Value)> {
-        vec![]
-    }
-}
