+++ abra_core/src/builtins/native/mod.rs
@@ -7,7 +7,7 @@ mod native_array;
 mod native_float;
 mod native_int;
 mod native_map;
-mod native_set;
+mod native_set2;
 mod native_string;
 
 pub use common::{NativeType, to_string, default_to_string_method};
@@ -16,5 +16,5 @@ pub use native_array::Array;
 pub use native_float::NativeFloat;
 pub use native_int::NativeInt;
 pub use native_map::Map;
-pub use native_set::NativeSet;
+pub use native_set2::Set;
 pub use native_string::NativeString;
