+++ abra_native/src/lib.rs
@@ -543,6 +543,11 @@ fn gen_rust_type_path(type_repr: &TypeRepr, type_ref_name: &String) -> proc_macr
             if type_ref_name == "Array" {
                 let arr_type_repr = TypeRepr::Array(Box::new(type_args[0].clone()));
                 gen_rust_type_path(&arr_type_repr, type_ref_name)
+            } else if type_ref_name == "Set" {
+                let inner_type = type_args.get(0).expect("Sets require T value");
+                let inner_type_repr = gen_rust_type_path(inner_type, type_ref_name);
+
+                quote! { crate::typechecker::types::Type::Set(std::boxed::Box::new(#inner_type_repr)) }
             } else if type_ref_name == "Map" {
                 let key_type = type_args.get(0).expect("Maps require K value");
                 let key_type_repr = gen_rust_type_path(key_type, type_ref_name);
