+++ abra_wasm/src/js_value/value.rs
@@ -90,13 +90,13 @@ impl<'a> Serialize for JsWrappedObjValue<'a> {
                 obj.serialize_entry("value", value)?;
                 obj.end()
             }
-            Obj::SetObj(value) => {
-                let mut obj = serializer.serialize_map(Some(2))?;
-                obj.serialize_entry("kind", "setObj")?;
-                let value: Vec<JsWrappedValue> = value.iter().map(|i| JsWrappedValue(i)).collect();
-                obj.serialize_entry("value", &value)?;
-                obj.end()
-            }
+            // Obj::SetObj(value) => {
+            //     let mut obj = serializer.serialize_map(Some(2))?;
+            //     obj.serialize_entry("kind", "setObj")?;
+            //     let value: Vec<JsWrappedValue> = value.iter().map(|i| JsWrappedValue(i)).collect();
+            //     obj.serialize_entry("value", &value)?;
+            //     obj.end()
+            // }
             Obj::TupleObj(value) => {
                 let mut obj = serializer.serialize_map(Some(2))?;
                 obj.serialize_entry("kind", "tupleObj")?;
